requirements.txt
appnope==0.1.3
asttokens==2.4.1
bidict==0.22.1
biosppy==1.0.0
colorama==0.4.6
colorlog==6.7.0
comm==0.2.0
contourpy==1.2.0
cramjam==2.7.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
easydev==0.12.1
exceptiongroup==1.1.3
executing==2.0.1
fastparquet==2023.10.1
fonttools==4.44.3
fsspec==2023.10.0
future==0.18.3
h5py==3.10.0
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
matplotlib==3.8.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
nolds==0.5.2
numpy==1.26.1
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.2
parso==0.8.3
pexpect==4.8.0
Pillow==10.1.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pyhrv==0.4.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.1
scikit-learn==1.3.2
scipy==1.11.3
shortuuid==1.0.11
six==1.16.0
spectrum==0.8.1
stack-data==0.6.3
threadpoolctl==3.2.0
tornado==6.3.3
traitlets==5.13.0
tzdata==2023.3
wcwidth==0.2.9
tqdm==4.66.1